refactor: 초기 위치와 관련된 클래스를 board 패키지로 이동
