+++ src/test/java/domain/board/wing/RightWingTest.java
@@ -1,7 +1,8 @@
-package domain.board;
+package domain.board.wing;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.board.Intersection;
 import domain.game.Side;
 import domain.piece.Piece;
 import domain.piece.PieceType;
@@ -10,27 +11,27 @@ import java.util.Map;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-class LeftWingTest {
+class RightWingTest {
 
     @Nested
-    class 좌진_내부_기물의_초기_위치를_반환한다 {
+    class 우진_내부_기물의_초기_위치를_반환한다 {
 
         @Test
         void 한_진영의_초기_위치를_반환한다() {
             // given
             Side side = Side.HAN;
-            LeftWing leftWing = new LeftWing(List.of(
+            RightWing rightWing = new RightWing(List.of(
                     new Piece(PieceType.HORSE, side),
                     new Piece(PieceType.ELEPHANT, side)
             ));
 
             Map<Intersection, Piece> expected = Map.of(
-                    new Intersection(1, 8), new Piece(PieceType.HORSE, side),
-                    new Intersection(1, 7), new Piece(PieceType.ELEPHANT, side)
+                    new Intersection(1, 3), new Piece(PieceType.HORSE, side),
+                    new Intersection(1, 2), new Piece(PieceType.ELEPHANT, side)
             );
 
             // when
-            Map<Intersection, Piece> actual = leftWing.setUpPieces(side);
+            Map<Intersection, Piece> actual = rightWing.setUpPieces(side);
 
             // then
             assertThat(actual).isEqualTo(expected);
@@ -40,18 +41,18 @@ class LeftWingTest {
         void 초_진영의_초기_위치를_반환한다() {
             // given
             Side side = Side.CHO;
-            LeftWing leftWing = new LeftWing(List.of(
+            RightWing rightWing = new RightWing(List.of(
                     new Piece(PieceType.HORSE, side),
                     new Piece(PieceType.ELEPHANT, side)
             ));
 
             Map<Intersection, Piece> expected = Map.of(
-                    new Intersection(10, 2), new Piece(PieceType.HORSE, side),
-                    new Intersection(10, 3), new Piece(PieceType.ELEPHANT, side)
+                    new Intersection(10, 7), new Piece(PieceType.HORSE, side),
+                    new Intersection(10, 8), new Piece(PieceType.ELEPHANT, side)
             );
 
             // when
-            Map<Intersection, Piece> actual = leftWing.setUpPieces(side);
+            Map<Intersection, Piece> actual = rightWing.setUpPieces(side);
 
             // then
             assertThat(actual).isEqualTo(expected);
